fox-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
pypdf
faiss-cpu
numpy
langchain
langchain_community
sentence-transformers
python-dotenv
httpx